// bconnect-universaldynamicgroups-mcp/src/generated/universaldynamicgroups-types.ts
export interface components {
    schemas: {
        Folder: {
            /**
             * Format: guid
             * @description The id of the folder
             */
            id?: string;
            /** @description The name of the folder */
            name?: string;
            /** @description The comment of the folder */
            comment?: string | null;
            /**
             * Format: guid
             * @description The id of the parent folder
             */
            parentId?: string;
            /** @description The name of the parent folder */
            parent?: string;
        };
        FolderPagedList: {
            /** Format: int32 */
            readonly currentPage?: number | null;
            /** Format: int32 */
            readonly pageSize?: number | null;
            /** Format: int32 */
            readonly totalPages?: number | null;
            /** Format: int32 */
            readonly totalItems?: number;
            readonly hasPreviousPage?: boolean;
            readonly hasNextPage?: boolean;
            readonly data?: components["schemas"]["Folder"][];
        };
        /** @description Array of operations to perform */
        JsonPatchDocument: components["schemas"]["Operation"][];
        Operation: {
            op?: string;
            value?: Record<string, never> | null;
            path?: string;
        };
        ProblemDetails: {
            type?: string | null;
            title?: string | null;
            /** Format: int32 */
            status?: number | null;
            detail?: string | null;
            instance?: string | null;
        } & {
            [key: string]: unknown;
        };
        UniversalDynamicGroup: {
            /**
             * Format: guid
             * @description The id of the universal dynamic group
             */
            id?: string;
            /** @description The name of the universal dynamic group */
            name?: string;
            /** @description The comment of the universal dynamic group */
            comment?: string | null;
            /** @description The name of the folder */
            folderName?: string;
            /**
             * Format: guid
             * @description The id of the folder
             */
            folderId?: string;
        };
        UniversalDynamicGroupPagedList: {
            /** Format: int32 */
            readonly currentPage?: number | null;
            /** Format: int32 */
            readonly pageSize?: number | null;
            /** Format: int32 */
            readonly totalPages?: number | null;
            /** Format: int32 */
            readonly totalItems?: number;
            readonly hasPreviousPage?: boolean;
            readonly hasNextPage?: boolean;
            readonly data?: components["schemas"]["UniversalDynamicGroup"][];
        };
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}
export type $defs = Record<string, never>;
export interface operations {
    GetUniversalDynamicGroups: {
        parameters: {
            query?: {
                /** @description Filters results by exactly matching name. */
                Name?: string;
                /** @description Sorts results by property name and with sort direction. Multiple properties are separated by comma. Possible values are Name or Comment (e. g. Name asc). */
                OrderBy?: string;
                /** @description Filters results by matching the given value against searchable properties. Searchable values are Name, Comment. */
                SearchQuery?: string;
                /** @description The zero-indexed number of the first page that begins the set of pages that are returned in the response. */
                Page?: number;
                /** @description The number of items to list on a single page. For example, if page_size is 10, each page shows ten items. <br />A valid value is a non-negative, non-zero integer. The default value is 20. Maximum value is 1000. */
                PageSize?: number;
            };
            header?: never;
            path?: never;
            cookie?: never;
        };
        requestBody?: never;
        responses: {
            /** @description Returns all universal dynamic groups */
            200: {
                headers: {
                    [name: string]: unknown;
                };
                content: {
                    "application/json": components["schemas"]["UniversalDynamicGroupPagedList"];
                };
            };
            /** @description Bad Request */
            400: {
                headers: {
                    [name: string]: unknown;
                };
                content: {
                    "application/json": components["schemas"]["ProblemDetails"];
                };
            };
            /** @description Internal Server Error */
            500: {
                headers: {
                    [name: string]: unknown;
                };
                content: {
                    "application/json": components["schemas"]["ProblemDetails"];
                };
            };
        };
    };
// ...
    GetFolders: {
        parameters: {
            query?: {
                /** @description Sorts results by property name and with sort direction. Multiple properties are separated by comma.<br />Possible value is Name. */
                OrderBy?: string;
                /** @description Filters results by matching the given value against searchable properties.<br />Searchable values are Name and Comment. */
                SearchQuery?: string;
                /** @description Filters result by matching the exact value against Name. */
                Name?: string;
                /** @description The zero-indexed number of the first page that begins the set of pages that are returned in the response. */
                Page?: number;
                /** @description The number of items to list on a single page. For example, if page_size is 10, each page shows ten items. <br />A valid value is a non-negative, non-zero integer. The default value is 20. Maximum value is 1000. */
                PageSize?: number;
            };
            header?: never;
            path?: never;
            cookie?: never;
        };
        requestBody?: never;
        responses: {
            /** @description Returns all folders */
            200: {
                headers: {
                    [name: string]: unknown;
                };
                content: {
                    "application/json": components["schemas"]["FolderPagedList"];
                };
            };
            /** @description Bad Request */
            400: {
                headers: {
                    [name: string]: unknown;
                };
                content: {
                    "application/json": components["schemas"]["ProblemDetails"];
                };
            };
            /** @description Internal Server Error */
            500: {
                headers: {
                    [name: string]: unknown;
                };
                content: {
                    "application/json": components["schemas"]["ProblemDetails"];
                };
            };
        };
    };
// ...
    GetFoldersByFolderId: {
        parameters: {
            query?: {
                /** @description Sorts results by property name and with sort direction. Multiple properties are separated by comma.<br />Possible value is Name. */
                OrderBy?: string;
                /** @description Filters results by matching the given value against searchable properties.<br />Searchable values are Name and Comment. */
                SearchQuery?: string;
                /** @description Filters result by matching the exact value against Name. */
                Name?: string;
                /** @description The zero-indexed number of the first page that begins the set of pages that are returned in the response. */
                Page?: number;
                /** @description The number of items to list on a single page. For example, if page_size is 10, each page shows ten items. <br />A valid value is a non-negative, non-zero integer. The default value is 20. Maximum value is 1000. */
                PageSize?: number;
                includeSubFolders?: boolean;
            };
            header?: never;
            path: {
                folderId: string;
            };
            cookie?: never;
        };
        requestBody?: never;
        responses: {
            /** @description Returns the folders contained in the specified folder */
            200: {
                headers: {
                    [name: string]: unknown;
                };
                content: {
                    "application/json": components["schemas"]["FolderPagedList"];
                };
            };
            /** @description Bad Request */
            400: {
                headers: {
                    [name: string]: unknown;
                };
                content: {
                    "application/json": components["schemas"]["ProblemDetails"];
                };
            };
            /** @description The specified folder cannot be found */
            404: {
                headers: {
                    [name: string]: unknown;
                };
                content: {
                    "application/json": components["schemas"]["ProblemDetails"];
                };
            };
            /** @description Internal Server Error */
            500: {
                headers: {
                    [name: string]: unknown;
                };
                content: {
                    "application/json": components["schemas"]["ProblemDetails"];
                };
            };
        };
    };
}